requirements.txt
fastapi>=0.104.1
pydantic>=2.5.0
redis>=5.0.0
argon2-cffi>=23.1.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
